refactor: refatorando validação, criado middleware
