+++ src/server/shared/middleware/Validation.ts
@@ -1,0 +1,23 @@
+import { RequestHandler } from "express";
+import { StatusCodes } from "http-status-codes";
+import { ObjectSchema, ValidationError } from "yup";
+
+type TValidation = (schema: ObjectSchema<any>) => RequestHandler;
+
+export const Validation: TValidation = (schema) => async (req, res, next) => {
+  try {
+    await schema.validate(req.body, {
+      abortEarly: false,
+    });
+    return next();
+  } catch (err) {
+    const yupError = err as ValidationError;
+    const errors: Record<string, string> = {};
+
+    yupError.inner.forEach((error) => {
+      if (!error.path) return;
+      errors[error.path] = error.message;
+    });
+    return res.status(StatusCodes.BAD_REQUEST).json({ errors });
+  }
+};
